customer_backend/requirements.txt
Django==3.2.11
django-environ==0.8.1
django-rest-auth==0.9.5
django-rest-auth[with_social]
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.3
